TimeseriesTools.AbstractTimeSeries

TimeseriesTools.BinaryTimeSeries

TimeseriesTools.IrregularIndex

TimeseriesTools.IrregularTimeIndex

TimeseriesTools.IrregularTimeSeries

TimeseriesTools.MultidimensionalTimeSeries

TimeseriesTools.MultivariateTimeSeries

TimeseriesTools.RegularIndex

TimeseriesTools.RegularTimeIndex

TimeseriesTools.RegularTimeSeries

TimeseriesTools.SpikeTrain

TimeseriesTools.TimeIndex

TimeseriesTools.UnivariateTimeSeries

TimeseriesTools.Var

TimeseriesTools.TimeSeries

TimeseriesTools.TimeSeries

TimeseriesTools.TimeSeries

TimeseriesTools.AbstractTimeSeries — Type

AbstractTimeSeries{T, N, B}

A type alias for an AbstractDimArray with a time index.

TimeseriesTools.BinaryTimeSeries — Type

BinaryTimeSeries

A type alias for a time series of bits.

TimeseriesTools.IrregularIndex — Type

IrregularIndex

A type alias for an irregularly sampled dimension, wrapping an AbstractVector.

TimeseriesTools.IrregularTimeIndex — Type

IrregularTimeIndex

A type alias for a tuple of dimensions containing a TimeIndex and any number of other dimensions.

TimeseriesTools.IrregularTimeSeries — Type

IrregularTimeSeries

A type alias for a potentially irregularly sampled time series.

TimeseriesTools.MultidimensionalTimeSeries — Type

A multidimensional time series has a regular sampling over a dimension other than time; a one-dimensional time series can be thought of as a field over an even grid in 1 dimension that fluctuates over time.

TimeseriesTools.MultivariateTimeSeries — Type

MultivariateTimeSeries{T}

A type alias for a multivariate time series (A matrix, with a first Ti dimension and an arbitrary second dimension).

TimeseriesTools.RegularIndex — Type

RegularIndex

A type alias for a regularly sampled dimension, wrapping an AbstractRange.

TimeseriesTools.RegularTimeIndex — Type

RegularTimeIndex

A type alias for a tuple of dimensions containing a TimeIndex and any number of other dimensions.

TimeseriesTools.RegularTimeSeries — Type

RegularTimeSeries{T, N, B}

A type alias for a regularly sampled time series.

TimeseriesTools.SpikeTrain — Type

SpikeTrain

A type alias for a spike-train time series, which contains spike times in the time dimension and true for all values corresponding to a spike. The spike times can be retrieved with times(x).

TimeseriesTools.TimeIndex — Type

TimeIndex

A type alias for a tuple containing a time dimension and any number of other dimensions.

TimeseriesTools.UnivariateTimeSeries — Type

UnivariateTimeSeries{T}

A type alias for a time series with one variable (a vector with only a Ti dimension).

TimeseriesTools.Var — Type

Var

A DimensionalData.jl dimension representing the variables of a multivariate time series.

TimeseriesTools.TimeSeries — Method

TimeSeries(t, x)

Constructs a univariate time series with time t and data x. Alteratively, use TS(t, x)

Examples

julia> using TimeseriesTools, Unitful;
julia> t = 1:100
julia> x = rand(100)
julia> ts = TimeSeries(t, x)
julia> ts isa typeintersect(UnivariateTimeSeries, RegularTimeSeries)

TimeseriesTools.TimeSeries — Method

TimeSeries(t, f::Function)

Construct a time series by mapping a function f over the time points t.

TimeseriesTools.TimeSeries — Method

TimeSeries(t, v, x)

Constructs a multivariate time series with time t, variable v, and data x.

Examples

julia> t = 1:100;
julia> v = [:a, :b, :c];
julia> x = rand(100, 3);
julia> mts = TimeSeries(t, v, x)
julia> mts isa typeintersect(MultivariateTimeSeries, RegularTimeSeries)