TimeseriesTools

Documentation for TimeseriesTools.

IntervalSets.Interval

TimeseriesTools.AbstractSpectrum

TimeseriesTools.AbstractTimeSeries

TimeseriesTools.Freq

TimeseriesTools.FreqIndex

TimeseriesTools.IrregularTimeSeries

TimeseriesTools.MultivariateSpectrum

TimeseriesTools.MultivariateTimeSeries

TimeseriesTools.RegularFreqIndex

TimeseriesTools.RegularIndex

TimeseriesTools.RegularSpectrum

TimeseriesTools.RegularTimeIndex

TimeseriesTools.RegularTimeSeries

TimeseriesTools.TimeIndex

TimeseriesTools.UnitfulIndex

TimeseriesTools.UnitfulTimeIndex

TimeseriesTools.UnitfulTimeSeries

TimeseriesTools.UnivariateTimeSeries

TimeseriesTools.Var

Base.step

TimeseriesTools.TimeSeries

TimeseriesTools.TimeSeries

TimeseriesTools._energyspectrum

TimeseriesTools._energyspectrum

TimeseriesTools._powerspectrum

TimeseriesTools.convertconst

TimeseriesTools.duration

TimeseriesTools.energyspectrum

TimeseriesTools.powerspectrum

TimeseriesTools.samplingperiod

TimeseriesTools.samplingrate

TimeseriesTools.times

TimeseriesTools.UnitfulIndex — Type

UnitfulIndex

A type alias for a union of AbstractArray, AbstractRange, and Tuple types with Unitful.Time elements.

IntervalSets.Interval — Method

IntervalSets.Interval(x::AbstractTimeSeries)

Returns an interval representing the range of the @ref x.

Examples

julia> using IntervalSets;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x);
julia> IntervalSets.Interval(ts) == (1..100)

TimeseriesTools.AbstractSpectrum — Type

AbstractSpectrum{T, N, B}

A type alias for an AbstractDimArray in which the first dimension is @refuency.

TimeseriesTools.AbstractTimeSeries — Type

AbstractTimeSeries{T, N, B}

A type alias for an AbstractDimArray with a time index.

TimeseriesTools.Freq — Type

Freq

A DimensionalData.jl dimension representing the frequency domain.

TimeseriesTools.FreqIndex — Type

FreqIndex

A type alias for a tuple of dimensions, where the first dimension is of type FreqDim.

TimeseriesTools.IrregularTimeSeries — Type

IrregularTimeSeries

A type alias for a potentially irregularly sampled time series.

TimeseriesTools.MultivariateSpectrum — Type

MultivariateSpectrum{T} = AbstractSpectrum{T, 2} where T

A type alias for a multivariate spectrum.

TimeseriesTools.MultivariateTimeSeries — Type

MultivariateTimeSeries{T}

A type alias for a multivariate time series (A matrix, with a first Ti dimension and an arbitrary second dimension).

TimeseriesTools.RegularFreqIndex — Type

RegularFreqIndex

A type alias for a tuple of dimensions, where the first dimension is a regularly sampled @refuency.

TimeseriesTools.RegularIndex — Type

RegularIndex

A type alias for a regularly sampled dimension, wrapping an AbstractRange.

TimeseriesTools.RegularSpectrum — Type

RegularSpectrum{T, N, B}

A type alias for a spectrum with a regularly sampled frequency index.

TimeseriesTools.RegularTimeIndex — Type

RegularTimeIndex

A type alias for a tuple of dimensions containing a @ref and any number of other dimensions.

TimeseriesTools.RegularTimeSeries — Type

RegularTimeSeries{T, N, B}

A type alias for a regularly sampled time series.

TimeseriesTools.TimeIndex — Type

TimeIndex

A type alias for a tuple containing a time dimension and any number of other dimensions.

TimeseriesTools.UnitfulTimeIndex — Type

UnitfulTimeIndex

A type alias for a tuple of dimensions, where the first dimension is of type DimensionalData.Dimension{<:UnitfulIndex}.

TimeseriesTools.UnitfulTimeSeries — Type

UnitfulTimeSeries{T, N, B}

A type alias for an AbstractDimArray with a UnitfulTimeIndex.

Examples

julia> using Unitful;
julia> t = (1:100)u"s";
julia> x = rand(100);
julia> uts = TimeSeries(t, x);
julia> uts isa UnitfulTimeSeries

TimeseriesTools.UnivariateTimeSeries — Type

UnivariateTimeSeries{T}

A type alias for a time series with one variable (a vector with only a Ti dimension).

TimeseriesTools.Var — Type

Var

A DimensionalData.jl dimension representing the variables of a multivariate time series.

Base.step — Method

step(x::RegularTimeSeries)

Returns the step size (time increment) of a regularly sampled @ref.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> rts = TimeSeries(t, x);
julia> step(rts) == 1

TimeseriesTools.TimeSeries — Method

TimeSeries(t, v, x)

Constructs a multivariate time series with time t, variable v, and data x.

Examples

julia> t = 1:100;
julia> v = [:a, :b, :c];
julia> x = rand(100, 3);
julia> mts = TimeSeries(t, v, x)
julia> mts isa Union{MultivariateTimeSeries, RegularTimeSeries}

TimeseriesTools.TimeSeries — Method

TimeSeries(t, x)

Constructs a univariate time series with time t and data x. Alteratively, use TS(t, x)

Examples

julia> t = 1:100
julia> x = rand(100)
julia> ts = TimeSeries(t, x)
julia> ts isa Union{UnivariateTimeSeries, RegularTimeSeries}

TimeseriesTools._energyspectrum — Function

_energyspectrum(x::RegularTimeSeries, f_min=samplingrate(x)/min(length(x)÷4, 1000))

Computes the energy spectrum of a regularly sampled time series x with an optional minimum frequency f_min.

TimeseriesTools._energyspectrum — Method

_energyspectrum(x::RegularTimeSeries, f_min=0)

Computes the energy spectrum of a time series using the fast Fourier transform.

If f_min > 0, the energy spectrum is calculated for windows of the time series determined by f_min, the minimum frequency that will be resolved in the spectrum. If f_min > 0, the second dimension of the output will correspond to the windows. For an averaged periodogram, see @ref.

If the input time series is a @ref, the frequency will also have units. Moreover if the elements of x are unitful, so are the elements of the spectrum.

Examples

```jldoctest julia> using TimeseriesTools julia> t = range(0.0, stop=1.0, length=1000); julia> x = sin.(2 * π * 5 * t); julia> ts = RegularTimeSeries(t, x); julia> S = _energyspectrum(ts); julia> S isa MultivariateSpectrum

TimeseriesTools._powerspectrum — Method

_powerspectrum(x::AbstractTimeSeries, f_min=samplingrate(x)/min(length(x)÷4, 1000); kwargs...)

Computes the power spectrum of a time series x in Welch periodogram windows. Note that the _powerspectrum is simply the @ref divided by the duration of each window. See @ref.

TimeseriesTools.convertconst — Method

TimeseriesTools.convertconst(c::Number, u::Unitful.Quantity)

Converts a constant c to have the same units as u.

Examples

julia> using Unitful;
julia> c = 5;
julia> u = 3u"s";
julia> converted_c = TimeseriesTools.convertconst(c, u);
julia> typeof(converted_c) == typeof(u)

TimeseriesTools.duration — Method

duration(x::AbstractTimeSeries)

Returns the duration of the @ref x.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x);
julia> duration(ts) == 99

TimeseriesTools.energyspectrum — Method

energyspectrum(x::RegularTimeSeries, f_min=0; kwargs...)

Computes the average energy spectrum of a regularly sampled time series x. f_min determines the minimum frequency that will be resolved in the spectrum. See @ref.

TimeseriesTools.powerspectrum — Method

powerspectrum(x::AbstractTimeSeries, f_minsamplingrate(x)/min(length(x)÷4, 1000); kwargs...)

Computes the average power spectrum of a time series x using the Welch periodogram method.

TimeseriesTools.samplingperiod — Method

samplingperiod(x::RegularTimeSeries)

Returns the sampling period (step size) of a regularly sampled @ref.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> rts = TimeSeries(t, x);
julia> samplingperiod(rts) == 1

TimeseriesTools.samplingrate — Method

samplingrate(x::RegularTimeSeries)

Returns the sampling rate (inverse of the step size) of a regularly sampled @ref.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> rts = TimeSeries(t, x);
julia> samplingrate(rts) == 1

TimeseriesTools.times — Method

times(x::AbstractTimeSeries)

Returns the time indices of the @ref x.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x);
julia> times(ts) == t