TimeseriesTools

Documentation for TimeseriesTools.

IntervalSets.Interval

TimeseriesTools.AbstractSpectrum

TimeseriesTools.AbstractTimeSeries

TimeseriesTools.BinaryTimeSeries

TimeseriesTools.Freq

TimeseriesTools.FreqIndex

TimeseriesTools.IrregularIndex

TimeseriesTools.IrregularTimeIndex

TimeseriesTools.IrregularTimeSeries

TimeseriesTools.MultivariateSpectrum

TimeseriesTools.MultivariateTimeSeries

TimeseriesTools.RegularFreqIndex

TimeseriesTools.RegularIndex

TimeseriesTools.RegularSpectrum

TimeseriesTools.RegularTimeIndex

TimeseriesTools.RegularTimeSeries

TimeseriesTools.TimeIndex

TimeseriesTools.UnitfulIndex

TimeseriesTools.UnitfulSpectrum

TimeseriesTools.UnitfulTimeIndex

TimeseriesTools.UnitfulTimeSeries

TimeseriesTools.UnivariateSpectrum

TimeseriesTools.UnivariateTimeSeries

TimeseriesTools.Var

Base.cat

Base.step

GeometryBasics.decompose

TimeseriesTools.Spectrum

TimeseriesTools.Spectrum

TimeseriesTools.TimeSeries

TimeseriesTools.TimeSeries

TimeseriesTools.TimeSeries

TimeseriesTools._energyspectrum

TimeseriesTools._energyspectrum

TimeseriesTools._powerspectrum

TimeseriesTools.closeneighbours

TimeseriesTools.colorednoise

TimeseriesTools.convertconst

TimeseriesTools.dimunit

TimeseriesTools.duration

TimeseriesTools.energyspectrum

TimeseriesTools.frequnit

TimeseriesTools.powerspectrum

TimeseriesTools.samplingperiod

TimeseriesTools.samplingrate

TimeseriesTools.stitch

TimeseriesTools.stoic

TimeseriesTools.sttc

TimeseriesTools.times

TimeseriesTools.timeunit

Unitful.unit

TimeseriesTools.UnitfulIndex — Type

UnitfulIndex

A type alias for a union of AbstractArray, AbstractRange, and Tuple types with Unitful.Time elements.

IntervalSets.Interval — Method

IntervalSets.Interval(x::AbstractTimeSeries)

Returns an interval representing the range of the AbstractTimeSeries x.

Examples

julia> using IntervalSets;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x);
julia> IntervalSets.Interval(ts) == (1..100)

TimeseriesTools.AbstractSpectrum — Type

AbstractSpectrum{T, N, B}

A type alias for an AbstractDimArray in which the first dimension is Frequency.

TimeseriesTools.AbstractTimeSeries — Type

AbstractTimeSeries{T, N, B}

A type alias for an AbstractDimArray with a time index.

TimeseriesTools.BinaryTimeSeries — Type

BinaryTimeSeries

A type alias for a time series of bits.

TimeseriesTools.Freq — Type

Freq

A DimensionalData.jl dimension representing the frequency domain.

TimeseriesTools.FreqIndex — Type

FreqIndex

A type alias for a tuple of dimensions, where the first dimension is of type FrequencyDim.

TimeseriesTools.IrregularIndex — Type

IrregularIndex

A type alias for an irregularly sampled dimension, wrapping an AbstractVector.

TimeseriesTools.IrregularTimeIndex — Type

IrregularTimeIndex

A type alias for a tuple of dimensions containing a TimeIndex and any number of other dimensions.

TimeseriesTools.IrregularTimeSeries — Type

IrregularTimeSeries

A type alias for a potentially irregularly sampled time series.

TimeseriesTools.MultivariateSpectrum — Type

MultivariateSpectrum{T} = AbstractSpectrum{T, 2} where T

A type alias for a multivariate spectrum.

TimeseriesTools.MultivariateTimeSeries — Type

MultivariateTimeSeries{T}

A type alias for a multivariate time series (A matrix, with a first Ti dimension and an arbitrary second dimension).

TimeseriesTools.RegularFreqIndex — Type

RegularFreqIndex

A type alias for a tuple of dimensions, where the first dimension is a regularly sampled Frequency.

TimeseriesTools.RegularIndex — Type

RegularIndex

A type alias for a regularly sampled dimension, wrapping an AbstractRange.

TimeseriesTools.RegularSpectrum — Type

RegularSpectrum{T, N, B}

A type alias for a spectrum with a regularly sampled frequency index.

TimeseriesTools.RegularTimeIndex — Type

RegularTimeIndex

A type alias for a tuple of dimensions containing a TimeIndex and any number of other dimensions.

TimeseriesTools.RegularTimeSeries — Type

RegularTimeSeries{T, N, B}

A type alias for a regularly sampled time series.

TimeseriesTools.TimeIndex — Type

TimeIndex

A type alias for a tuple containing a time dimension and any number of other dimensions.

TimeseriesTools.UnitfulSpectrum — Type

UnitfulSpectrum{T,N,B}

A type representing spectra with unitful frequency units.

TimeseriesTools.UnitfulTimeIndex — Type

UnitfulTimeIndex

A type alias for a tuple of dimensions, where the first dimension is of type DimensionalData.Dimension{<:UnitfulIndex}.

TimeseriesTools.UnitfulTimeSeries — Type

UnitfulTimeSeries{T, N, B}

A type alias for an AbstractDimArray with a UnitfulTimeIndex.

Examples

julia> using Unitful;
julia> t = (1:100)u"s";
julia> x = rand(100);
julia> uts = TimeSeries(t, x);
julia> uts isa UnitfulTimeSeries

TimeseriesTools.UnivariateSpectrum — Type

UnivariateSpectrum{T} = AbstractSpectrum{T, 1} where T

A type alias for a univariate spectrum.

TimeseriesTools.UnivariateTimeSeries — Type

UnivariateTimeSeries{T}

A type alias for a time series with one variable (a vector with only a Ti dimension).

TimeseriesTools.Var — Type

Var

A DimensionalData.jl dimension representing the variables of a multivariate time series.

Base.cat — Method

Base.cat(D::DimensionalData.Dimension, args...; kwargs...)

Concatenate the arrays given in args..., and give the resulting extra axis dimensions D. Note that unlike Base.cat without the first Dim argument, this increments all existing dimensions greater than dims by one (so N n×n arrays concatenated at dims=1 will result in an N×n×n array).

Base.step — Method

step(x::RegularTimeSeries)

Returns the step size (time increment) of a regularly sampled RegularTimeSeries.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> rts = TimeSeries(t, x);
julia> step(rts) == 1

GeometryBasics.decompose — Method

decompose(x::Union{<:AbstractTimeSeries, <:AbstractSpectrum})

Convert a time series or spectrum to a tuple of the dimensions and the data (as Arrays).

TimeseriesTools.Spectrum — Method

Spectrum(f, v, x)

Constructs a multivariate spectrum with frequencies f, variables v, and data x.

TimeseriesTools.Spectrum — Method

Spectrum(f, x)

Constructs a univariate time series with frequencies f and data x.

TimeseriesTools.TimeSeries — Method

TimeSeries(t, v, x)

Constructs a multivariate time series with time t, variable v, and data x.

Examples

julia> t = 1:100;
julia> v = [:a, :b, :c];
julia> x = rand(100, 3);
julia> mts = TimeSeries(t, v, x)
julia> mts isa typeintersect(MultivariateTimeSeries, RegularTimeSeries)

TimeseriesTools.TimeSeries — Method

TimeSeries(t, x, timeunit::Unitful.Units)

Constructs a univariate time series with time t, data x and time units specified by timeunit. Note that you can add units to the elements of a time series x with, for example, x*u"V".

Examples

julia> using Unitful;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x, u"ms")*u"V";
julia> ts isa Union{UnivariateTimeSeries, RegularTimeSeries, UnitfulTimeSeries}

TimeseriesTools.TimeSeries — Method

TimeSeries(t, x)

Constructs a univariate time series with time t and data x. Alteratively, use TS(t, x)

Examples

julia> using TimeseriesTools, Unitful;
julia> t = 1:100
julia> x = rand(100)
julia> ts = TimeSeries(t, x)
julia> ts isa typeintersect(UnivariateTimeSeries, RegularTimeSeries)

TimeseriesTools._energyspectrum — Function

_energyspectrum(x::RegularTimeSeries, f_min=samplingrate(x)/min(length(x)÷4, 1000))

Computes the energy spectrum of a regularly sampled time series x with an optional minimum frequency f_min.

TimeseriesTools._energyspectrum — Method

_energyspectrum(x::RegularTimeSeries, f_min=0)

Computes the energy spectrum of a time series using the fast Fourier transform.

If f_min > 0, the energy spectrum is calculated for windows of the time series determined by f_min, the minimum frequency that will be resolved in the spectrum. If f_min > 0, the second dimension of the output will correspond to the windows. For an averaged periodogram, see energyspectrum.

If the input time series is a UnitfulTimeSeries, the frequency will also have units. Moreover if the elements of x are unitful, so are the elements of the spectrum.

Examples

julia> using TimeseriesTools
julia> t = range(0.0, stop=1.0, length=1000);
julia> x = sin.(2 * π * 5 * t);
julia> ts = RegularTimeSeries(t, x);
julia> S = _energyspectrum(ts);
julia> S isa MultivariateSpectrum

TimeseriesTools._powerspectrum — Method

_powerspectrum(x::AbstractTimeSeries, f_min=samplingrate(x)/min(length(x)÷4, 1000); kwargs...)

Computes the power spectrum of a time series x in Welch periodogram windows. Note that the _powerspectrum is simply the _energyspectrum divided by the duration of each window. See _energyspectrum.

TimeseriesTools.closeneighbours — Method

closeneighbours(x, y; Δt)

Constructs a sparse matrix of distances between neighbouring spikes in two sorted spike trains.

Arguments

x: A sorted array representing the first spike train.

y: A sorted array representing the second spike train.

Δt: The maximum time difference allowed for two spikes to be considered neighbours.

Returns

A sparse matrix D where D[i, j] represents the distance between the i-th spike in x and the j-th spike in y, for pairs of spikes within Δt of each other.

TimeseriesTools.colorednoise — Method

colorednoise(ts::AbstractRange; α=2.0)

Generate a colored-noise time series with a specified power-law exponent α on the given times ts.

Arguments

ts: An AbstractRange representing the time range of the generated noise.

α: The power-law exponent of the colored noise, which will have a spectrum given by 1/f^α. Defaults to 2.0.

Returns

A TimeSeries containing the generated colored noise.

Example

julia> using TimeseriesTools
julia> pink_noise = colorednoise(1:0.01:10; α=1.0)
julia> pink_noise isa RegularTimeSeries

TimeseriesTools.convertconst — Method

TimeseriesTools.convertconst(c::Number, u::Unitful.Quantity)

Converts a constant c to have the same units as u.

Examples

julia> using Unitful;
julia> c = 5;
julia> u = 3u"s";
julia> converted_c = TimeseriesTools.convertconst(c, u);
julia> typeof(converted_c) == typeof(u)

TimeseriesTools.dimunit — Method

dimunit(x::UnitfulTimeSeries, dim)

Returns the unit associated with the specified dimension dim of a UnitfulTimeSeries.

Examples

julia> using Unitful;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x, u"ms");
julia> TimeseriesTools.dimunit(ts, Ti) == u"ms"

TimeseriesTools.duration — Method

duration(x::AbstractTimeSeries)

Returns the duration of the AbstractTimeSeries x.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x);
julia> TimeseriesTools.duration(ts) == 99

TimeseriesTools.energyspectrum — Method

energyspectrum(x::RegularTimeSeries, f_min=0; kwargs...)

Computes the average energy spectrum of a regularly sampled time series x. f_min determines the minimum frequency that will be resolved in the spectrum. See _energyspectrum.

TimeseriesTools.frequnit — Method

frequnit(x::UnitfulSpectrum)

Returns the frequency units associated with a UnitfulSpectrum.

Examples

julia> using Unitful;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x, u"ms");
julia> sp = fft(ts);  # assuming fft returns a UnitfulSpectrum
julia> frequnits(sp) == u"Hz"

TimeseriesTools.powerspectrum — Method

powerspectrum(x::AbstractTimeSeries, f_min=samplingrate(x)/min(length(x)÷4, 1000); kwargs...)

Computes the average power spectrum of a time series x using the Welch periodogram method.

TimeseriesTools.samplingperiod — Method

samplingperiod(x::RegularTimeSeries)

Returns the sampling period (step size) of a regularly sampled RegularTimeSeries.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> rts = TimeSeries(t, x);
julia> samplingperiod(rts) == 1

TimeseriesTools.samplingrate — Method

samplingrate(x::RegularTimeSeries)

Returns the sampling rate (inverse of the step size) of a regularly sampled RegularTimeSeries.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> rts = TimeSeries(t, x);
julia> samplingrate(rts) == 1

TimeseriesTools.stitch — Method

stitch(x, args...)

Stitch multiple time series together by concatenatign along the time dimension generating new contiguous time indices. The time series must be of the same type (UnivariateRegular, MultivariateRegular, or AbstractArray), and the sampling period and dimensions of the data arrays must match. If the arguments are `MultivariateRegular, they must have the same dimensions (except for the time dimension).

Arguments

X: The first time series.

args...: Additional time series.

Returns

A new time series containing the concatenated data.

TimeseriesTools.stoic — Method

stoic(a, b; kpi = npi, σ = 0.025, Δt = σ * 10)

Compute the spike-train overlap-integral covariance between two spike trains, after normalizing both convolutions to unit energy

Arguments

a: Spike train a.

b: Spike train b.

kpi: Kernel product integral, a function of the distance between two spikes. Default is npi, the integral of two gaussians with equal variance at a given distance from each other.

σ: Width parameter of the kernel. For npi, this is the width of the unit-mass Gaussian kernels. Default is 0.025.

Δt: Time window for considering spikes as close. Default is σ * 10.

Returns

The stoichiometry between spike trains a and b.

TimeseriesTools.sttc — Method

sttc(a, b; Δt = 0.025)

The spike-time tiling coefficient, a measure of correlation between spike trains [1].

Arguments

a::Vector{<:Real}: A sorted vector of spike times.

b::Vector{<:Real}: A second sorted vector of spike times .

Δt::Real=0.025: The time window for calculating the STTC.

Returns

sttc::Real: The STTC value.

References

[1] [Cutts & Eglen 2014](https://doi.org/10.1523%2FJNEUROSCI.2767-14.2014)

TimeseriesTools.times — Method

times(x::AbstractTimeSeries)

Returns the time indices of the AbstractTimeSeries x.

Examples

julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x);
julia> times(ts) == t

TimeseriesTools.timeunit — Method

timeunit(x::UnitfulTimeSeries)

Returns the time units associated with a [UnitfulTimeSeries].

Examples

julia> using Unitful;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x, u"ms");
julia> timeunit(ts) == u"ms"

Unitful.unit — Method

unit(x::AbstractArray)

Returns the units associated with the elements of an UnitfulTimeSeries or UnitfulSpectrum.

Examples

julia> using Unitful;
julia> t = 1:100;
julia> x = rand(100);
julia> ts = TimeSeries(t, x, u"ms")*u"V";
julia> unit(ts) == u"V"