TimeSeriesTools

Documentation for TimeSeriesTools.

TimeSeriesTools.TimeDelayEmbedding

TimeSeriesTools.get_reasonable_params

TimeSeriesTools.semivariogram

TimeSeriesTools.times

TimeSeriesTools.TimeDelayEmbedding — Method

TimeDelayEmbedding(Z::AbstractVector; nrow=100, method=:Forward)

NOTE: The resulting matrix treats each column in the returned matrix H as an embedded featured vector.

TimeSeriesTools.get_reasonable_params — Method

get_reasonable_params(γ,h)

Given an empirical semivariogram γ with time lags h, return reasonable default parameters for fitting a variogram model using the ParameterHandling format with positive constraints. Nugget is defaulted to 0.01 of maximum γ, partialsill is set to 85% of max γ value, and range is set to 10% of maximum lag.

TimeSeriesTools.semivariogram — Method

semivariogram(Z::TimeSeriesTools.AbstractRegularTimeSeries; lag_ratio=0.5, lag_max=100.0)

Provided a regular time series Z, compute the empirical semivariogram γ for time lags ranging from the minimum time step Δt to a maximum value. The maximum time lag is taken to be the smaller of lag_ratio times the full duration of the time series and lag_max, a user defined maximum lag time in the time units of Z.

Output

Returns a tuple (γ, h) containing the (scaled) variances and time lags.

TimeSeriesTools.times — Method

times(series::AbstractGenericTimeSeries)

Given a timeseries series return a vector of times