TimeSeriesTools

Documentation for TimeSeriesTools.

TimeSeriesTools.fit_spherical_γ

TimeSeriesTools.semivariogram

TimeSeriesTools.times

TimeSeriesTools.fit_spherical_γ — Method

fit_spherical_γ(h, γ, params; optargs...)

Given an empirical semi-variogram γ with lags h, fit a spherical variogram model to the data with parameters params. The arguments optargs are passed to the optimize function.

TimeSeriesTools.semivariogram — Method

semivariogram(Z::TimeSeriesTools.AbstractRegularTimeSeries; lag_ratio=0.5, lag_max=100.0)

Provided a regular time series Z, compute the empirical semivariogram γ for time lags ranging from the minimum time step Δt to a maximum value. The maximum time lag is taken to be the smaller of lag_ratio times the full duration of the time series and lag_max, a user defined maximum lag time in the time units of Z.

Output

Returns a tuple (γ, h) containing the (scaled) variances and time lags.

TimeSeriesTools.times — Method

times(series::AbstractGenericTimeSeries)

Given a timeseries series return a vector of times